requirements_gemini.txt
Flask==2.3.3
pandas==2.0.3
mlxtend==0.23.0
numpy==1.24.3
Pillow==10.0.0
opencv-python==4.8.0.76
python-dotenv==1.0.0
google-generativeai==0.3.2